move settings saving to subclass meant just for MMStudio
